Created Organization of MDX Files
I organized the files using a
chapter/a/*.mdx file. This file will
contain all the problems for the subsection.
I chose against making unique files for each
problem because I want to use anchor tags
on the website which will scroll down to the
problems and use the sidebar to navigate to them.